Move is_portable_alias_name to yash_env::alias
Make the portable-name check available to the alias built-in and other
crates that depend on yash-env, without going through yash-syntax. The
parser still reaches it via the yash_syntax::alias re-export.
